Updated repository interfaces and documentation
The IDatabaseRepository<TEntity, TType> and IDatabasesRepository<TEntity> interfaces have been updated, adding support for CancellationToken to all methods and new DeleteAsync and UpdateRangeAsync methods. The documentation of the new methods and code examples in the readme.md file have also been updated. In addition, the implementation of the IDatabaseRepository<TEntity, TType> interface in the DatabaseRepository<TEntity, TType> class has been updated according to the new changes.