Updated version and corrected conditions in DatabaseRepository
Updated the Assembly version in Directory.Packages.props from 0.1.2 to 0.1.3. Changed the condition that checks the result of the `SaveChangesAsync` method in the `DatabaseRepository` class. Now instead of checking if the result is greater than 0, it checks if the result is equal to 0. This change applies to the `DeleteAsync` and `Update` methods. In addition, comments about throwing a `DatabaseException` exception when there are problems with updating or deleting records have been removed.